+++ cmd/config.go
@@ -14,4 +14,5 @@ func init() {
 	RootCmd.AddCommand(configCmd)
 	persistentFlags := configCmd.PersistentFlags()
 	persistentFlags.StringVarP(&ospreyconfigFile, "ospreyconfig", "o", "", "osprey targets configuration. Defaults to $HOME/.osprey/config")
+	persistentFlags.StringVarP(&targetGroup, "group", "g", "", "show only the specified group")
 }